+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 
 #### Bug fixes
  - only show ipv4 in UI and in avahi aliases.
+ - fixed experiment profile plugin checks
+ - fixed experiment profile display crashing the UI when editing plugins
 
 #### Breaking changes
  - `pio clear-cache` renamed to `pio cache clear`